requirements.txt
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
diffusers==0.16.1
fastapi==0.95.1
uvicorn==0.12.0
transformers>=4.25.1
google-cloud-storage
matplotlib
torch>=1.4
torchvision
accelerate
firebase-admin